dbprov-master/pyproject.toml
[project]
name = "dbprov"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Matej Gallo", email = "422328@mail.muni.cz"},
]
dependencies = []
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
prov-gen = [
    "prov>=2.0.1",
    "psycopg2-binary>=2.9.10",
    "pony>=0.7.19",
    "pydot>=3.0.2",
    "tqdm>=4.66.6",
    "mlflow>=2.17.2",
    "dirhash>=0.5.0",
    "cryptography>=43.0.3",
    "jcs>=0.2.1",
]
prov-storage = [
    "prov>=2.0.1",
    "django-neomodel>=0.2.0",
    "requests>=2.32.3",
    "django-health-check>=3.18.3",
]
trusted-party = [
    "prov>=2.0.1",
    "psycopg2-binary>=2.9.10",
    "jcs>=0.2.1",
    "pyopenssl>=24.2.1",
    "Django>=5.1.3",
    "django-health-check>=3.18.3",
]
[tool.pdm]
distribution = false
